i2c-imx: do not allow interruptions when waiting for I2C to complete
X-Git-Tag: v2.6.36~10^2
X-Git-Url: http://opensource.freescale.com/git?p=imx%2Flinux-2.6-imx.git;a=commitdiff_plain;h=e39428d53d080ad2615b772d7f99d2a70c2aaab2

The i2c_imx_trx_complete() function is using
wait_event_interruptible_timeout() to wait for the I2C controller to
signal that it has completed an I2C bus operation. If the process that
causes the I2C operation receives a signal, the wait will be
interrupted, returning an error. It is better to let the I2C operation
finished before handling the signal (i.e. returning into userspace).

It is safe to use wait_event_timeout() instead, because the timeout
will allow the process to exit if the I2C bus hangs. It's also better
to allow the I2C operation to finish, because unacknowledged I2C
operations can cause the I2C bus to hang.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>
Reviewed-by: Wolfram Sang <w.sang@pengutronix.de>
Signed-off-by: Ben Dooks <ben-linux@fluff.org>
Signed-off-by: Markus Niebel <Markus.Niebel@tqs.de>
Signed-off-by: Stephan Linz <stephan.linz@clondiag.com>