MTD: cfi_cmdset_0002 - toggle bit workaround for M29EW
The erase/program controller on Numonyx M29EW NOR flash sometimes
does not work as written in the datasheet. Two sequential status reads
give the same result but erase, program or suspend are not ready yet.
This causes wrong read results during erase suspend and the possibility
of file system crashs.

The Probability of the error raises with temperature. Always call
chip_good when erasing / programming to test for operation complete and
verify suspend ready with additional query to DQ7 in active erasing
sector.

Signed-off-by: Markus Niebel <markus.niebel@tqs.de>
Signed-off-by: Stephan Linz <stephan.linz@clondiag.com>