ARM: mach-mx3: tqma35 - fix size of EEPROM on TQMa35 board
The EEPROM is optional. If available, it is always an M24C64 wich has 64 kiByte (8192 x 8 Bits)
Give the correct value to enable the full memory size. Also give the correct device name "24c64"
to better reflect the hardware in sysfs and help the driver to deal with misconfiguration etc.

Signed-off-by: Markus Niebel <Markus.Niebel@tqs.de>
Signed-off-by: Stephan Linz <stephan.linz@clondiag.com>